requirements.txt
Flask==1.1.2
Jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==0.16.1
Flask-SQLAlchemy==2.4.4
Flask-Bcrypt==1.0.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.9
flask-restplus==0.13.0
Flask-Testing==0.8.1
requests==2.25.0
mysqlclient==2.0.1
pika==1.1.0
python-dotenv
SQLAlchemy==1.3.20